// Server/src/Module/Building/Building.Enum.ts
export enum EBuildingType {
  None = 0,
  Castle = 101,
  Wall = 102,
  ArrowTower = 103,
  WatchTower = 104,
  Warehouse = 105,
  EventCenter = 106,
  Merchant = 107,
  Port = 108,
  TroopDetails = 109,
  Bulletin = 110,
  Embassy = 111,
  Market = 112,
  Institute = 113,
  TreasurePool = 114,
  Blacksmith = 115,
  Stable = 116,
  TargetRange = 117,
  Barrack = 118,
  ChariotPlant = 119,
  Fortress = 120,
  DrillGrounds = 121,
  HallOfWar = 122,
  TrainHall = 123,
  Prison = 125,
  Farm = 201,
  LumberMill = 202,
  IronMine = 203,
  SilverMine = 204,
  MarchingTent = 205,
  FirstAidTent = 206,
  CrystalMine = 207,
  MaterialWorkShop = 210,
  Miracle = 301,
  ElitePalace = 302,
  Monument = 303,
  StarBraveStatue = 304,
  LeisureHouse = 30001,
  HIP = 30002,
  PetHouse = 30003,
  ResurrectionHall = 30004,
  ServiceCenter = 30005,
  EpicBattle = 30006

}

export enum EBuildingPos {
  CBPlace_None = 0,
  CBPlace_Fixed = 1,
  CBPlace_Inner = 2,
  CBPlace_Outer = 3,
};

export enum EWatchTowerEffect{

}